requirements.txt
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
applicationinsights==0.11.10
argcomplete==1.12.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
azure-cli-core==2.35.0
azure-cli-telemetry==1.0.6
azure-common==1.1.28
azure-core==1.23.1
azure-identity==1.9.0
azure-mgmt-authorization==2.0.0
azure-mgmt-compute==26.1.0
azure-mgmt-confluent==1.0.0
azure-mgmt-containerservice==18.0.0
azure-mgmt-core==1.3.0
azure-mgmt-msi==6.0.0
azure-mgmt-network==19.3.0
azure-mgmt-resource==21.0.0
azure-mgmt-storage==20.0.0
backoff==1.11.1
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
frozenlist==1.3.0
google-api-core==2.7.1
google-auth==2.6.3
googleapis-common-protos==1.56.0
humanfriendly==10.0
idna==3.3
isodate==0.6.1
jmespath==1.0.0
knack==0.9.0
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.2
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
packaging==21.3
paramiko==2.10.3
pep8==1.7.1
pkginfo==1.8.2
portalocker==1.7.1
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tabulate==0.8.9
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
